chore: Add init container to download scannercli
The scannercli is no longer added to the adapters's Docker image.
Instead, it is mounted at /usr/local/bin/scannercli through volumes.

For Kubernetes we configure the init container to download the scannercli
binary from the Aqua registry and then store it in a shared volume,
which is finally mounted at /usr/local/bin/scannercli and visible
to the main container.

Resolves: #16

Signed-off-by: Daniel Pacak <pacak.daniel@gmail.com>