fix(scannercli): Pass temporary robot account credentials to scannercli (#50)
Pass Robot Account credentials generated by Harbor for each scan
request through to the scannercli command as --robot-username
and --robot-password flags.

This feature is enabled by setting the value of the
SCANNER_CLI_OVERRIDE_REGISTRY_CREDENTIALS to true.

It has to be enable explicitly as not every version of scannercli
accepts the --robot-username and --robot-password flags.

Resolves: #37

Co-authored-by: Simarpreet Singh <simar@linux.com>
Signed-off-by: Daniel Pacak <pacak.daniel@gmail.com>